+++ include/keyboard.h
@@ -3,8 +3,13 @@
 
 #include <stdint.h>
 
+#define KEY_UP    0x80
+#define KEY_DOWN  0x81
+#define KEY_LEFT  0x82
+#define KEY_RIGHT 0x83
+
 void keyboard_init(void);
-char keyboard_getchar(void);
+unsigned char keyboard_getchar(void);
 void keyboard_wait_for_key(void);
 
 #endif